+++ Assets/Scripts/MVZ2/UI/Level/EnergyPanel.cs
@@ -5,10 +5,16 @@ namespace MVZ2.Level.UI
 {
     public class EnergyPanel : LevelUIUnit
     {
+        public void FlickerEnergy()
+        {
+            textAnimator.SetTrigger("Flicker");
+        }
         public void SetEnergy(string energy)
         {
             energyText.text = energy;
         }
+        [SerializeField]
+        private Animator textAnimator;
         [SerializeField]
         private TextMeshProUGUI energyText;
     }